+++ app/services/constants.js
@@ -1,9 +1,5 @@
-const constants = Object.freeze({
-  REDIS: {
-    AUTH_USER_KEY: 'user_auth',
-    DATA_USAGE_KEY: 'user_usage_data',
-    AUTH_DATE_KEY: 'user_auth_date'
-  }
+export const REDIS = Object.freeze({
+  AUTH_USER_KEY: 'user_auth',
+  DATA_USAGE_KEY: 'user_usage_data',
+  AUTH_DATE_KEY: 'user_auth_date'
 })
-
-export const REDIS = constants.REDIS
