--- telegram-bot/constants/redis.js
@@ -1,7 +1,0 @@
-module.exports = Object.freeze({
-  AUTH_USER_KEY: 'user_auth',
-  DATA_USAGE_KEY: 'user_usage_data',
-  AUTH_DATE_KEY: 'user_auth_date',
-  ADMIN_USER_KEY: 'user_admin',
-  USER_STATE: 'user_state'
-})
