--- telegram-bot/constants/user-state.js
@@ -1,6 +1,0 @@
-module.exports = Object.freeze({
-  IDLE: 'idle',
-  CREATE_USER_ENTER_USERNAME: 'create_user_enter_username',
-  CREATE_USER_ENTER_PASSWORD: 'create_user_enter_password',
-  DELETE_USER_ENTER_USERNAME: 'delete_user_enter_username'
-})
