+++ telegram-bot/package.json
@@ -2,7 +2,8 @@
   "name": "telegram-bot",
   "version": "1.5.0",
   "description": "",
-  "main": "app.js",
+  "type": "module",
+  "exports": "./app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "nodemon app.js",
