--- telegram-bot/services/redis/index.js
@@ -1,7 +1,0 @@
-const redis = require('redis')
-
-module.exports = () => redis.createClient({
-  host: process.env.REDIS_HOST,
-  port: process.env.REDIS_PORT,
-  db: process.env.REDIS_DB
-})
